Add category module
