+++ src/modules/categories/types/category.ts
@@ -1,0 +1,4 @@
+export interface Category {
+  uuid: string;
+  name: string;
+}
